requirements.txt
# Smart CLM - Updated Requirements for Python 3.13
# Install with: pip install -r requirements.txt

# FastAPI Framework - Latest versions
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Latest Pydantic that supports Python 3.13
pydantic==2.10.3
pydantic[email]==2.10.3
python-multipart==0.0.12

# Templates
jinja2==3.1.4

# Security (for production)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
bcrypt==4.2.1

# File handling
aiofiles==24.1.0

# Development
httpx==0.27.2
pytest==8.3.4
pytest-asyncio==0.24.0

# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.1

# Pydantic
pydantic==2.10.3
pydantic[email]==2.10.3
pydantic-settings==2.3.4

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-dotenv==1.0.1

# Redis (for caching/sessions)
redis==5.2.0
hiredis==2.3.2

# Email
python-multipart==0.0.12
emails==0.6

# File handling
aiofiles==24.1.0
python-magic==0.4.27

# Templates
jinja2==3.1.4

# Development
httpx==0.27.2
pytest==8.3.4
pytest-asyncio==0.24.0

# Database migrations
alembic==1.13.1

# Logging
loguru==0.7.2

# Date handling
python-dateutil==2.9.0.post0

fastapi-mail==1.4.1


# Optional Database drivers (uncomment as needed)
# sqlalchemy==2.0.36
# psycopg2-binary==2.9.10  # PostgreSQL
# motor==3.6.0  # MongoDB async driver
# redis==5.2.0  # Redis for sessions

#correspondence
httpx==0.27.0
anthropic==0.18.1

PyPDF2>=3.0.1
python-docx>=0.8.11

anthropic==0.39.0

reportlab
apscheduler==3.10.4
twilio==8.10.0
boto3==1.28.85

